requirements.txt
pandas~=1.2.3
Quandoo~=1.3.7
python-dateutil~=2.8.1
python-dotenv~=0.15.0